Add async audio decoder.
AudioContext would use AsyncAudioDecoder to decode audio data.
AsyncAudioDecoder has its own thread to do the actual decoding. It tries
to create an AudioFileReader to read the audio data and store
the audio data into the AudioBuffer. After decoding is done, we run the
success callback if decoding succeed, otherwise run the error callback if
it exists.

In this CL, we support WAV reading which reads 32 bit floats and 16 bit
PCM.

b/21713250

+++ cobalt/audio/audio_node.h
@@ -64,10 +64,14 @@ class AudioNode : public dom::EventTarget {
 
   // The number of inputs feeding into the AudioNode. For source nodes, this
   // will be 0.
-  size_t number_of_inputs() const { return inputs_.size(); }
+  uint32 number_of_inputs() const {
+    return static_cast<uint32>(inputs_.size());
+  }
   // The number of outputs coming out of the AudioNode. This will be 0 for an
   // AudioDestinationNode.
-  size_t number_of_outputs() const { return outputs_.size(); }
+  uint32 number_of_outputs() const {
+    return static_cast<uint32>(outputs_.size());
+  }
 
   // The number of channels used when up-mixing and down-mixing connections to
   // any inputs to the node. The default value is 2 except for specific nodes
